+++ pika/src/main/java/com/numlock/pika/service/ReviewService.java
@@ -8,6 +8,6 @@ import java.util.List;
 public interface ReviewService {
     ReviewResponseDto createReview(ReviewRequestDto reviewRequestDto);
     ReviewResponseDto getReviewById(Long reviewId);
-    List<ReviewResponseDto> getReviewsByProductId(Long productId);
+    List<ReviewResponseDto> getReviewsByProductId(int productId);
     // Potentially add methods for update and delete later
 }
